Crate kerria_runtime

Re-exports

pub use constants::currency;

pub use constants::fee;

pub use constants::paras;

pub use constants::time;

pub use impls::DealWithFees;

pub use pallet_liquid_staking;

pub use pallet_amm;

pub use pallet_bridge;

pub use pallet_crowdloans;

pub use pallet_farming;

pub use pallet_loans;

pub use pallet_nominee_election;

pub use pallet_prices;

pub use pallet_router;

pub use frame_support::log;

Modules

api

constants

impls

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AMMPalletId

AcaPerSecond

AccountIdToMultiLocation

AggregatedDataProvider

Ancestry

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetAccountDeposit

AssetDeposit

AssetsStringLimit

AusdPerSecond

BaseCallFilter

BaseXcmWeight

BasicDeposit

BlockExecutionWeight

Importing a block with 0 txs takes ~5 ms

BlockHashCount

BondingDuration

BridgeMaxMembers

BridgePalletId

Burn

CallFilterRouter

ChangeBridgeMembers

CooloffPeriod

CrowdloansMigrationV1

CrowdloansPalletId

CurrencyIdConvert

Decimal

DefaultLpFee

DefaultProtocolFee

DefaultProtocolFeeReceiver

DepositBase

DepositFactor

DerivativeIndex

DotPerSecond

EnactmentPeriod

EraLength

ExistentialDeposit

ExpiresIn

ExtrinsicBaseWeight

Executing 10,000 System remarks (no-op) txs takes ~1.26 seconds -> ~125 µs per tx

FarmingPalletId

FastTrackVotingPeriod

FieldDeposit

GeneralCouncilMaxMembers

GeneralCouncilMaxProposals

GeneralCouncilMotionDuration

GenesisConfig

GiftAccount

GiftConvert

IdentityFee

Implementor of WeightToFeePolynomial that maps one unit of weight to one unit of fee.

InstantAllowed

KerriaNetwork

LCDOTPerSecond

LDOTPerSecond

LaunchPeriod

LiquidCurrency

LiquidStakingAgentsMembershipMaxMembers

LoansPalletId

LockPoolMaxDuration

MaxAdditionalFields

MaxApprovals

MaxAssetsForTransfer

MaxAuthorities

MaxCandidates

MaxHasDispatchedSize

MaxInstructions

MaxInvulnerables

MaxLengthRoute

MaxLocks

MaxPending

MaxProposals

MaxProxies

MaxRegistrars

MaxRewardTokens

MaxScheduledPerBlock

MaxSignatories

MaxSubAccounts

MaxUserLockItemsCount

MaxValidators

MaxVestingSchedules

MaxVotes

MaxVrfs

MaximumSchedulerWeight

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MigrateKeysLimit

MinCandidates

MinContribution

MinStake

MinUnstake

MinVestedTransfer

MinimumCount

MinimumDeposit

MinimumLiquidity

MinimumPeriod

NativeCurrencyId

NoPreimagePostponement

NotifyTimeout

Offset

OneAccount

OperationalFeeMultiplier

OracleMaxMembers

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaPerSecond

ParaPerSecondOfCanonicalLocation

ParallelKerria

ParentOrSiblings

Period

PotId

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMaximum

ProposalBondMinimum

ProposalLifetime

ProxyDepositBase

ProxyDepositFactor

RefundLocation

RelayChainBlockNumberProvider

RelayChainOrigin

RelayCurrency

RelayLocation

RelayNetwork

RemoveKeysLimit

ReservedDmpWeight

ReservedXcmpWeight

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

RouterPalletId

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SchedulerMigrationV3

SelfLocation

SpendPeriod

StakingCurrency

StakingPalletId

SubAccountDeposit

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

ThresholdPercentage

ToTreasury

TransactionByteFee

TreasuryAccount

TreasuryPalletId

UncleGenerations

Version

VotingPeriod

WhiteListFilter

XDOTPerSecond

XDOTPerSecondOfCanonicalLocation

XcmConfig

XcmFees

XcmHelperPalletId

Enums

Call

DispatchClass

A generalized group of dispatch types.

Event

OriginCaller

ProxyType

XCMPMessage

Constants

NATIVE_ASSET_ID

PRIMARY_PROBABILITY

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

InstanceFilter

Simple trait for providing a filter over a reference to some type, given an instance of itself.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AMM

AMMRoute

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Assets

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Barrier

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Bridge

BridgeMembership

BridgeMembershipConfig

CheckedExtrinsic

Extrinsic type that has already been checked.

CollatorSelection

CollatorSelectionConfig

Crowdloans

CumulusXcm

CurrencyAdapter

Democracy

DemocracyConfig

DmpQueue

EmergencyShutdown

Executive

Executive: handles dispatch to the various modules.

Farming

GeneralCouncil

GeneralCouncilConfig

GeneralCouncilMembership

GeneralCouncilMembershipConfig

Header

Block header type as expected by this runtime.

Identity

LiquidStaking

LiquidStakingAgentsMembership

LiquidStakingAgentsMembershipConfig

LiquidStakingConfig

Loans

LocalAssetTransactor

Means for transacting assets on this chain.

LocalOriginToLocation

Local origins on this chain are allowed to dispatch XCM sends/executions. However, we later block this via ExecuteXcmOrigin.

LocationToAccountId

Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

Multisig

NomineeElection

Oracle

OracleMembership

OracleMembershipConfig

OrmlXcm

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Preimage

Prices

Proxy

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalCommitteeMembership

TechnicalCommitteeMembershipConfig

TimeStampedPrice

Timestamp

Trader

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig

XTokens

XcmHelper

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.

XcmpQueue